% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_events.R
\name{get_combination_df}
\alias{get_combination_df}
\title{Function to Get the Future Df}
\usage{
get_combination_df(new_data, val, date_col, model, events_df, regressors_df)
}
\arguments{
\item{new_data}{Data frame or tibble}

\item{val}{Value from get_combination_value()}

\item{date_col}{Date column name}

\item{model}{Model}

\item{events_df}{Data frame with the events}

\item{regressors_df}{Data frame with the regressors}
}
\description{
Function to Get the Future Df
}
