% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_events.R
\name{get_combination_value}
\alias{get_combination_value}
\title{Function to Get a Number with the Combination Value}
\usage{
get_combination_value(args)
}
\arguments{
\item{args}{List with the arguments passed to the function.}
}
\description{
Function to Get a Number with the Combination Value
}
