% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-install.R
\name{install_nprophet}
\alias{install_nprophet}
\title{Install Neural Prophet}
\usage{
install_nprophet()
}
\description{
Installs \verb{Neural Prophet} Deep Learning Time Series Forecasting Software
using \code{reticulate::py_install()}.
\itemize{
\item A \code{Python} Environment will be created
named \code{nprophet}.
\item The neuralprophet R package will connect to the \code{nprophet} Python environment
}
}
\details{
\strong{Options for Connecting to Python}
\itemize{
\item \strong{Recommended} \emph{Use Pre-Configured Python Environment:} Use \code{install_nprophet()} to
install Neural Prophet Python Libraries into a conda environment named 'nprophet'.
\item \strong{Advanced} \emph{Use a Custom Python Environment:} Before running \code{library(neuralprohet)},
use \code{Sys.setenv(NPROPHET_PYTHON = 'path/to/python')} to set the path of your
python executable in an environment that has 'pillow', 'torch', 'numpy', 'pandas',
and 'neuralprophet' available as dependencies.
}

\strong{Package Manager Support (Python Environment)}
\itemize{
\item \strong{Conda Environments:} Currently, \code{install_nprophet()} supports Conda and Miniconda Environments.
\item \strong{Virtual Environments:} are not currently supported with the default installation method, \code{install_nprophet()}.
However, you can connect to virtual environment that you have created using
\code{Sys.setenv(NPROPHET_PYTHON = 'path/to/python')} prior to running \code{library(neuralprohet)}.
}
}
\examples{
\dontrun{
install_nprophet()
}


}
