% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-neuralprophet.R
\name{neural_prophet_predict_impl}
\alias{neural_prophet_predict_impl}
\title{Bridge prediction Function for NEURAL PROPHET Models}
\usage{
neural_prophet_predict_impl(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Additional arguments passed to \code{neuralprophet::NeuralProphet()}}
}
\description{
Bridge prediction Function for NEURAL PROPHET Models
}
