% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nprophet-env.R
\name{nprophet-env}
\alias{nprophet-env}
\alias{is_nprophet_activated}
\alias{activate_nprophet}
\alias{get_python_env}
\alias{check_nprophet_dependencies}
\alias{detect_default_nprophet_env}
\title{Neural Prophet Environment Connection Utilities}
\usage{
is_nprophet_activated()

activate_nprophet()

get_python_env()

check_nprophet_dependencies()

detect_default_nprophet_env()
}
\description{
These functions are used for testing and establishing a python environment
connection with \code{neuralprophet}.
}
\details{
\itemize{
\item \code{is_nprophet_activated()}: Determines if a Neural Prophet Environment has been activated
during \code{library(neuralprophet)}.
\itemize{
\item If \code{TRUE}, then you should be good to go.
\item If \code{FALSE}, then a connection between \code{neuralprophet}
and your neuralprophet Python has \emph{not} been activated.
}
\item \code{activate_nprophet()}: Attempts to activate a connection between \code{neuralprophet}
and an associated NeuralProphet Python Environment using \code{reticulate::use_condaenv(required = TRUE)}.
\itemize{
\item It first looks for the system environment variable, 'NPROPHET_PYTHON', for a path to the python executable
\item It next looks for a Conda Environment named 'nprophet' (this is what most users will have)
}
\item \code{get_python_env()}: Returns the configuration for the python environment that is being discovered
using \code{reticulate::py_discover_config()}.
\item \code{check_nprophet_dependencies()}: Checks whether Neural Prophet required python dependencies are present in the
currently activated Python Environment.
\item \code{detect_default_nprophet_env()}: Detects if an 'nprophet' python environment is available.
\itemize{
\item Returns a \code{tibble} containing the
\item Returns \code{NULL} if an 'nprophet' environment is not detected
}
}
}
\examples{
\donttest{
# Returns TRUE if NeuralProphet connection established on package load
is_nprophet_activated()

#

}



}
\seealso{
\itemize{
\item \code{\link[=install_nprophet]{install_nprophet()}} - Used to install the python environment needed to run \code{neuralprophet}.
}
}
