% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-neuralprophet_params.R
\name{nprophet_params}
\alias{nprophet_params}
\alias{trend_reg}
\alias{trend_reg_threshold}
\alias{num_hidden_layers}
\alias{d_hidden}
\alias{ar_sparsity}
\title{Tuning Parameters for Neural Prophet Models}
\usage{
trend_reg(range = c(0, 100), trans = NULL)

trend_reg_threshold(range = c(0, 10), trans = NULL)

num_hidden_layers(range = c(0L, 10L), trans = NULL)

d_hidden(range = c(0L, 500L), trans = NULL)

ar_sparsity(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Tuning Parameters for Neural Prophet Models
}
\details{
The main parameters for Neural Prophet models are:
\itemize{
\item \code{trend_reg}: the trend rate changes can be regularized by setting trend_reg to a value greater zero.
This is a useful feature that can be used to automatically detect relevant changepoints.
\item \code{trend_reg_threshold}: Threshold for the trend regularization
\item \code{num_hidden_layers}: num_hidden_layers defines the number of hidden layers of the FFNNs used in the overall model.
\item \code{d_hidden}: d_hidden is the number of units in the hidden layers.
\item \code{ar_sparsity}: For ar_sparsity values in the range 0-1 are expected with 0 inducing complete sparsity and 1 imposing no regularization at
all
}
}
\examples{
trend_reg()

num_hidden_layers()

ar_sparsity()


}
