% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_events.R
\name{split_convert}
\alias{split_convert}
\title{Function to Convert events}
\usage{
split_convert(split_df)
}
\arguments{
\item{split_df}{Data frame or tibble}
}
\description{
Function to Convert events
}
