% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_events.R
\name{split_events}
\alias{split_events}
\title{Function to split events}
\usage{
split_events(df, pattern = "events")
}
\arguments{
\item{df}{Data frame or tibble}

\item{pattern}{Search pattern}
}
\description{
Function to split events
}
